can run some java code

+++ src/com/mogujie/wangzhi/jvmgo/inst/compare/if_icmp.go
@@ -13,7 +13,7 @@ func (self *IF_ICMPEQ) Execute(frame *run.Frame) {
 	v2 := frame.OpStack().PopInt()
 	v1 := frame.OpStack().PopInt()
 	if v1 == v2 {
-		base.Branch(frame, self.Offset)
+		base.Branch(frame, int32(self.Offset))
 	}
 }
 
@@ -25,7 +25,7 @@ func (self *IF_ICMPNE) Execute(frame *run.Frame) {
 	v2 := frame.OpStack().PopInt()
 	v1 := frame.OpStack().PopInt()
 	if v1 != v2 {
-		base.Branch(frame, self.Offset)
+		base.Branch(frame, int32(self.Offset))
 	}
 }
 
@@ -37,7 +37,7 @@ func (self *IF_ICMPLT) Execute(frame *run.Frame) {
 	v2 := frame.OpStack().PopInt()
 	v1 := frame.OpStack().PopInt()
 	if v1 < v2 {
-		base.Branch(frame, self.Offset)
+		base.Branch(frame, int32(self.Offset))
 	}
 }
 
@@ -49,7 +49,7 @@ func (self *IF_ICMPLE) Execute(frame *run.Frame) {
 	v2 := frame.OpStack().PopInt()
 	v1 := frame.OpStack().PopInt()
 	if v1 <= v2 {
-		base.Branch(frame, self.Offset)
+		base.Branch(frame, int32(self.Offset))
 	}
 }
 
@@ -61,7 +61,7 @@ func (self *IF_ICMPGT) Execute(frame *run.Frame) {
 	v2 := frame.OpStack().PopInt()
 	v1 := frame.OpStack().PopInt()
 	if v1 > v2 {
-		base.Branch(frame, self.Offset)
+		base.Branch(frame, int32(self.Offset))
 	}
 }
 
@@ -73,6 +73,6 @@ func (self *IF_ICMPGE) Execute(frame *run.Frame) {
 	v2 := frame.OpStack().PopInt()
 	v1 := frame.OpStack().PopInt()
 	if v1 >= v2 {
-		base.Branch(frame, self.Offset)
+		base.Branch(frame, int32(self.Offset))
 	}
 }
